+++ src/tiled/changelayer.h
@@ -84,7 +84,7 @@ class SetLayerOpacity : public QUndoCommand
 public:
     SetLayerOpacity(MapDocument *mapDocument,
                     Layer *layer,
-                    float opacity);
+                    qreal opacity);
 
     void undo() override { setOpacity(mOldOpacity); }
     void redo() override { setOpacity(mNewOpacity); }
@@ -94,12 +94,12 @@ public:
     bool mergeWith(const QUndoCommand *other) override;
 
 private:
-    void setOpacity(float opacity);
+    void setOpacity(qreal opacity);
 
     MapDocument *mMapDocument;
     Layer *mLayer;
-    float mOldOpacity;
-    float mNewOpacity;
+    qreal mOldOpacity;
+    qreal mNewOpacity;
 };
 
 /**
